# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "YOLO Life Timer API",
    "description": "API documentation for YOLO Life Timer Backend MVP",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/life/configLife": {
      "post": {
        "tags": ["Life"],
        "summary": "Configure Life Parameters",
        "operationId": "configLife",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifeConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/life/getLifeConfig": {
      "get": {
        "tags": ["Life"],
        "summary": "Get Life Configuration",
        "operationId": "getLifeConfig",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLifeConfigDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/life/getLifeStatus": {
      "get": {
        "tags": ["Life"],
        "summary": "Get Life Status",
        "operationId": "getLifeStatus",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLifeStatusDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/goals": {
      "get": {
        "tags": ["Goal"],
        "summary": "Get All Goals",
        "operationId": "getGoals",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by status (e.g. 'ARCHIVED')"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListGoalWithStatsDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Goal"],
        "summary": "Create or Update Goal",
        "operationId": "saveGoal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGoal"
                }
              }
            }
          }
        }
      }
    },
    "/api/goals/{goalId}": {
      "get": {
        "tags": ["Goal"],
        "summary": "Get Goal By ID",
        "operationId": "getGoal",
        "parameters": [
          {
            "name": "goalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGoal"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Goal"],
        "summary": "Delete Goal",
        "operationId": "deleteGoal",
        "parameters": [
          {
            "name": "goalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/focus/start": {
      "post": {
        "tags": ["Focus"],
        "summary": "Start Focus Session",
        "operationId": "startFocus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFocusSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/focus/end": {
      "post": {
        "tags": ["Focus"],
        "summary": "End Focus Session",
        "operationId": "endFocus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusEndRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFocusSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/focus/running": {
      "get": {
        "tags": ["Focus"],
        "summary": "Get Running Focus Session",
        "operationId": "getRunningSession",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "goalId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFocusSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/focus": {
      "get": {
        "tags": ["Focus"],
        "summary": "Get Focus Session List",
        "operationId": "getFocusList",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "goalId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListFocusSessionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/focus/{id}": {
      "get": {
        "tags": ["Focus"],
        "summary": "Get Focus Session By ID",
        "operationId": "getFocusById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFocusSessionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/focus/statistics": {
      "get": {
        "tags": ["Focus"],
        "summary": "Get Focus Statistics",
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "goalId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFocusStatsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/milestones": {
      "get": {
        "tags": ["Milestone"],
        "summary": "Get Milestone List",
        "operationId": "getMilestoneList",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "goalId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListMilestone"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Milestone"],
        "summary": "Create Milestone",
        "operationId": "saveMilestone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMilestone"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Milestone"],
        "summary": "Update Milestone",
        "operationId": "updateMilestone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMilestone"
                }
              }
            }
          }
        }
      }
    },
    "/api/milestones/{id}": {
      "get": {
        "tags": ["Milestone"],
        "summary": "Get Milestone By ID",
        "operationId": "getMilestoneById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMilestone"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Milestone"],
        "summary": "Delete Milestone",
        "operationId": "deleteMilestone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "ApiResponseLifeStatusDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/LifeStatusDto"
          }
        }
      },
      "ApiResponseLifeConfigDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/LifeConfigDto"
          }
        }
      },
      "ApiResponseGoal": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/Goal"
          }
        }
      },
...
      "ApiResponseListGoalWithStatsDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalWithStatsDto"
            }
          }
        }
      },
      "ApiResponseFocusSession": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/FocusSession"
          }
        }
      },
      "ApiResponseFocusStatsDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/FocusStatsDto"
          }
        }
      },
      "LifeConfigRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 12345
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "example": "1995-05-20"
          },
          "expectedLifeYears": {
            "type": "integer",
            "example": 80
          },
          "energyLifeYears": {
            "type": "integer",
            "example": 60
          }
        }
      },
      "LifeStatusDto": {
        "type": "object",
        "properties": {
          "totalDays": {
            "type": "integer",
            "format": "int64"
          },
          "usedDays": {
            "type": "integer",
            "format": "int64"
          },
          "energyDays": {
            "type": "integer",
            "format": "int64"
          },
          "usedRatio": {
            "type": "string"
          },
          "lifeClock": {
            "type": "string"
          }
        }
      },
      "LifeConfigDto": {
        "type": "object",
        "properties": {
          "birthDate": {
            "type": "string",
            "format": "date",
            "example": "1995-05-20"
          },
          "expectedLifeYears": {
            "type": "integer",
            "example": 80
          },
          "energyLifeYears": {
            "type": "integer",
            "example": 60
          }
        }
      },
      "GoalSaveRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 12345
          },
          "title": {
            "type": "string",
            "example": "Learn Java"
          },
          "description": {
            "type": "string",
            "example": "Master Spring Boot"
          },
          "expectedTotalHours": {
            "type": "integer",
            "example": 100
          },
          "northStar": {
            "type": "string",
            "example": "Become a Senior Java Developer"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "description": "Status: ACTIVE, COMPLETED, ARCHIVED"
          }
        }
      },
      "Goal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "expectedTotalHours": {
            "type": "integer"
          },
          "northStar": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GoalWithStatsDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Goal"
          },
          {
            "type": "object",
            "properties": {
              "last7DaysMinutes": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ]
      },
      "FocusStartRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 12345
          },
          "goalId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      },
      "FocusEndRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 12345
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "durationMinutes": {
            "type": "integer"
          },
          "memo": {
            "type": "string"
          }
        }
      },
      "FocusSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "goalId": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "durationMinutes": {
            "type": "integer"
          },
          "memo": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FocusSessionDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FocusSession"
          },
          {
            "type": "object",
            "properties": {
              "goalTitle": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ApiResponseListFocusSessionDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FocusSessionDto"
            }
          }
        }
      },
      "ApiResponseFocusSessionDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/FocusSessionDto"
          }
        }
      },
      "FocusStatsDto": {
        "type": "object",
        "properties": {
          "goalId": {
            "type": "integer",
            "format": "int64"
          },
          "goalTitle": {
            "type": "string"
          },
          "last7DaysMinutes": {
            "type": "integer",
            "format": "int64"
          },
          "last30DaysMinutes": {
            "type": "integer",
            "format": "int64"
          },
          "dailyRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyFocusRecord"
            }
          }
        }
      },
      "DailyFocusRecord": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "minutes": {
            "type": "integer"
          }
        }
      },
      "Milestone": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "goalId": {
            "type": "integer",
            "format": "int64"
          },
          "milestoneTitle": {
            "type": "string"
          },
          "milestoneDate": {
            "type": "string",
            "format": "date"
          },
          "milestoneDesc": {
            "type": "string"
          },
          "ownFeel": {
            "type": "string"
          },
          "createAt": {
            "type": "string",
            "format": "date-time"
          },
          "updateAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MilestoneRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 12345
          },
          "goalId": {
            "type": "integer",
            "format": "int64"
          },
          "milestoneTitle": {
            "type": "string"
          },
          "milestoneDate": {
            "type": "string",
            "format": "date"
          },
          "milestoneDesc": {
            "type": "string"
          },
          "ownFeel": {
            "type": "string"
          }
        }
      },
      "ApiResponseMilestone": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/Milestone"
          }
        }
      },
      "ApiResponseListMilestone": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Milestone"
            }
          }
        }
      },
...
    }
  }
}