# src/main/java/com/life/yolo/controller/LifeController.java
package com.life.yolo.controller;

import com.life.yolo.common.ApiResponse;
import com.life.yolo.service.LifeService;
import lombok.Data;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.time.LocalDate;

@RestController
@RequestMapping("/api/life")
public class LifeController {

    @Autowired
    private LifeService lifeService;

    @PostMapping("/configLife")
    public ApiResponse<Void> configLife(@RequestBody LifeConfigRequest request) {
        lifeService.configureLife(request.getUserId(), request.getBirthDate(), request.getExpectedLifeYears(), request.getEnergyLifeYears());
        return ApiResponse.success(null);
    }

    @GetMapping("/getLifeConfig")
    public ApiResponse<LifeService.LifeConfigDto> getLifeConfig(@RequestParam Long userId) {
        return ApiResponse.success(lifeService.getLifeConfig(userId));
    }

    @GetMapping("/getLifeStatus")
    public ApiResponse<LifeService.LifeStatusDto> getLifeStatus(@RequestParam Long userId) {
        return ApiResponse.success(lifeService.getLifeStatus(userId));
    }

    @Data
    public static class LifeConfigRequest {
        private Long userId;
        private LocalDate birthDate;
        private Integer expectedLifeYears;
        private Integer energyLifeYears;
    }
}